Agregados reportes Top Clientes, Top Proveedores y Productos más vendidos

+++ app/models/compra.py
@@ -6,8 +6,8 @@ class Compra(db.Model):
     proveedor_id = db.Column(db.Integer, db.ForeignKey('proveedor.id'), nullable=False)
     producto_id = db.Column(db.Integer, db.ForeignKey('producto.id'), nullable=False)
     cantidad = db.Column(db.Integer, nullable=False)
-    precio_unitario = db.Column(db.Float, nullable=False)
-    total = db.Column(db.Float, nullable=False)
+    precio_unitario = db.Column(db.Integer, nullable=False)
+    total = db.Column(db.Integer, nullable=False)
 
     proveedor = db.relationship('Proveedor', backref='compras')
     producto = db.relationship('Producto', backref='compras')
